# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JABEEN Investor Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects (scoped by role)
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, agreement number, sector, investor name/company
        - name: status
          in: query
          schema:
            type: string
            enum: [on-track, delayed, stalled, complete]
        - name: stage
          in: query
          schema:
            type: string
          description: Filter by current stage name
        - name: sector
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project (managers only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project detail
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project metadata (managers only)
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project (admin only)
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "204":
          description: Deleted

  # --- STATUS UPDATES ---
...
  /projects/{projectId}/notes:
    get:
      operationId: listInternalNotes
      tags: [notes]
      summary: List internal notes (managers only)
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InternalNote"
    post:
      operationId: createInternalNote
      tags: [notes]
      summary: Create internal note (managers only)
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalNote"

  # --- NOTIFICATIONS ---
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    User:
      type: object
      required: [id, email, fullName, companyName, role, active, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        companyName:
          type: string
        title:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [investor, top-management, project-manager, administrator]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    ProjectSummary:
      type: object
      required: [id, name, sector, agreementNumber, derivedStatus, constructionPct, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        agreementNumber:
          type: string
        plotNumber:
          type: ["string", "null"]
        derivedStatus:
          type: string
          enum: [on-track, delayed, stalled, complete]
        constructionPct:
          type: integer
        attentionFlag:
          type: boolean
        lastUpdateAt:
          type: ["string", "null"]
          format: date-time
        currentStage:
          $ref: "#/components/schemas/StageSummary"
        investor:
          $ref: "#/components/schemas/InvestorContact"
        pipelineName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    Project:
      type: object
      required: [id, name, sector, agreementNumber, derivedStatus, constructionPct, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        agreementNumber:
          type: string
        plotNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        derivedStatus:
          type: string
          enum: [on-track, delayed, stalled, complete]
        constructionPct:
          type: integer
        attentionFlag:
          type: boolean
        lastUpdateAt:
          type: ["string", "null"]
          format: date-time
        currentStage:
          $ref: "#/components/schemas/StageSummary"
        investor:
          $ref: "#/components/schemas/InvestorContact"
        pipeline:
          $ref: "#/components/schemas/Template"
        pipelineId:
          type: ["integer", "null"]
        currentStageId:
          type: ["integer", "null"]
        investorId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name, sector, agreementNumber]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        sector:
          type: string
          minLength: 1
          maxLength: 200
        agreementNumber:
          type: string
          minLength: 1
          maxLength: 100
        plotNumber:
          type: string
          maxLength: 100
        notes:
          type: string
        pipelineId:
          type: integer
        constructionPct:
          type: integer
          minimum: 0
          maximum: 100
        investorId:
          type: integer
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        sector:
          type: string
          minLength: 1
          maxLength: 200
        plotNumber:
          type: ["string", "null"]
          maxLength: 100
        notes:
          type: ["string", "null"]
        attentionFlag:
          type: boolean
        constructionPct:
          type: integer
          minimum: 0
          maximum: 100
        investorId:
          type: ["integer", "null"]
        pipelineId:
          type: ["integer", "null"]
    StageSummary:
      type: object
      required: [id, name, orderIndex, progressBaseline, category]
      properties:
        id:
          type: integer
        name:
          type: string
        orderIndex:
          type: integer
        progressBaseline:
          type: integer
        category:
          type: string
          enum: [on-hold, active, complete]
        description:
          type: ["string", "null"]
    InvestorContact:
      type: object
      required: [id, fullName, email, companyName]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        companyName:
          type: string
        phone:
          type: ["string", "null"]

    # --- STATUS UPDATE ---
...
    InternalNote:
      type: object
      required: [id, projectId, authorId, body, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 4000

    # --- NOTIFICATION ---
...
    Template:
      type: object
      required: [id, name, isDefault, stages, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isDefault:
          type: boolean
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
        createdAt:
          type: string
          format: date-time
...
    Stage:
      type: object
      required: [id, name, orderIndex, progressBaseline, category, fields]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        orderIndex:
          type: integer
        progressBaseline:
          type: integer
          minimum: 0
          maximum: 100
        category:
          type: string
          enum: [on-hold, active, complete]
        fields:
          type: array
          items:
            $ref: "#/components/schemas/StageField"
...
    StageField:
      type: object
      required: [id, name, baseType, widget, required, position]
      properties:
        id:
          type: integer
        name:
          type: string
        baseType:
          type: string
          enum: [text, number, date, boolean, file, image, single-choice, multi-choice]
        widget:
          type: string
          enum: [single-line, multi-line, email, telephone, number, date, checkbox, toggle, file-upload, single-photo, photo-gallery, drop-list, list-box, radio, checkbox-list]
        required:
          type: boolean
        position:
          type: integer
        options:
          type: array
          items:
            type: string
        config:
          type: object
...