# lib/api-zod/src/generated/types/internalNote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JABEEN Investor Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface InternalNote {
  id: number;
  projectId: number;
  authorId: number;
  author?: User;
  body: string;
  createdAt: Date;
}